Move LinearAlgebra.hpp back to simulators/lightning_qubit but keep randomUnitary in TestHelpers.
