+++ pennylane_lightning/src/simulators/lightning_kokkos/tests/Test_StateVectorLKokkos.cpp
@@ -20,7 +20,6 @@
 
 #include <catch2/catch.hpp>
 
-#include "LinearAlgebra.hpp" //randomUnitary
 #include "StateVectorKokkos.hpp"
 #include "TestHelpers.hpp" // createRandomStateVectorData
