+++ pennylane_lightning/src/simulators/lightning_qubit/tests/Test_StateVectorLQubit.cpp
@@ -11,13 +11,6 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include "LinearAlgebra.hpp" //randomUnitary
-#include "StateVectorLQubitManaged.hpp"
-#include "StateVectorLQubitRaw.hpp"
-#include "TestHelpers.hpp" // createRandomStateVectorData
-#include "TestHelpersWires.hpp"
-#include "cpu_kernels/GateImplementationsPI.hpp"
-
 #include <algorithm>
 #include <complex>
 #include <limits> // numeric_limits
@@ -26,6 +19,13 @@
 #include <vector>
 
 #include <catch2/catch.hpp>
+
+#include "LinearAlgebra.hpp" //randomUnitary
+#include "StateVectorLQubitManaged.hpp"
+#include "StateVectorLQubitRaw.hpp"
+#include "TestHelpers.hpp" // createRandomStateVectorData
+#include "TestHelpersWires.hpp"
+#include "cpu_kernels/GateImplementationsPI.hpp"
 
 /**
  * @file
